added dependency injection in app layer
