+++ lib/data/dataSource/news_remote_data_source_impl.dart
@@ -31,7 +31,3 @@ class NewsRemoteDataSourceImpl extends NewsRemoteDataSource {
     }
   }
 }
-
-NewsRemoteDataSource injectNewsDataSource() {
-  return NewsRemoteDataSourceImpl(ApiManager.getInstance());
-}
