+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:news/presentation/home/home_screen.dart';
 import 'package:news/presentation/news/news_details_screen.dart';
 
-import 'my_theme.dart';
+import 'presentation/style/my_theme.dart';
 
 void main() {
   runApp(const MyApp());
