--- lib/presentation/category/di.dart
@@ -1,6 +1,0 @@
-import 'package:news/data/repository/sources_repository_impl.dart';
-import 'package:news/domain/usecases/getSourcesByCategoryIdUseCase.dart';
-
-GetSourcesByCategoryIdUseCase injectGetSourcesByCategoryIdUseCase() {
-  return GetSourcesByCategoryIdUseCase(injectSourcesRepository());
-}
