--- lib/presentation/news/di.dart
@@ -1,6 +1,0 @@
-import 'package:news/data/repository/news_repository_impl.dart';
-import 'package:news/domain/usecases/getNewsUseCase.dart';
-
-GetNewsUseCase injectGetNewsUseCase() {
-  return GetNewsUseCase(injectNewsRepository());
-}
